Rename loki.build to loki.jit_build
