+++ loki/backend/tests/test_pygen.py
@@ -14,7 +14,7 @@ import numpy as np
 
 from loki import Subroutine
 from loki.backend import pygen
-from loki.build import jit_compile, clean_test
+from loki.jit_build import jit_compile, clean_test
 from loki.frontend import available_frontends, OMNI
 from loki.transformations.transpile import FortranPythonTransformation
