+++ loki/jit_build/lib.py
@@ -17,9 +17,9 @@ except ImportError:
 
 from loki.logging import warning
 from loki.tools import as_tuple, find_paths
-from loki.build.compiler import _default_compiler
-from loki.build.obj import Obj
-from loki.build.workqueue import workqueue, wait_and_check
+from loki.jit_build.compiler import _default_compiler
+from loki.jit_build.obj import Obj
+from loki.jit_build.workqueue import workqueue, wait_and_check
 
 
 __all__ = ['Lib']
