+++ loki/jit_build/tests/test_build.py
@@ -8,8 +8,8 @@
 from pathlib import Path
 import pytest
 
-from loki.build import Obj, Lib, Builder
-from loki.build.compiler import  (
+from loki.jit_build import Obj, Lib, Builder
+from loki.jit_build.compiler import  (
     Compiler, GNUCompiler, NvidiaCompiler, get_compiler_from_env,
     _default_compiler
 )
