+++ loki/transformations/tests/test_transform_loop.py
@@ -11,7 +11,7 @@ import pytest
 import numpy as np
 
 from loki import Subroutine
-from loki.build import jit_compile, clean_test
+from loki.jit_build import jit_compile, clean_test
 from loki.expression import symbols as sym
 from loki.frontend import available_frontends
 from loki.ir import (
