+++ loki/transformations/transpile/tests/test_sdfg.py
@@ -12,7 +12,7 @@ import numpy as np
 import pytest
 
 from loki import Subroutine
-from loki.build import jit_compile
+from loki.jit_build import jit_compile
 from loki.frontend import available_frontends
 
 from loki.transformations.transpile import FortranPythonTransformation
